+++ src/v1/function-configuration.ts
@@ -215,7 +215,7 @@ export interface RuntimeOptions {
   /**
    * Specific service account for the function to run as.
    */
-  serviceAccount?: "default" | string | ResetValue;
+  serviceAccount?: "default" | string | Expression<string> | ResetValue;
 
   /**
    * Ingress settings which control where this function can be called from.
@@ -248,6 +248,30 @@ export interface RuntimeOptions {
   enforceAppCheck?: boolean;
 
   /**
+   * Determines whether Firebase App Check token is consumed on request. Defaults to false.
+   *
+   * @remarks
+   * Set this to true to enable the App Check replay protection feature by consuming the App Check token on callable
+   * request. Tokens that are found to be already consumed will have request.app.alreadyConsumed property set true.
+   *
+   *
+   * Tokens are only considered to be consumed if it is sent to the App Check service by setting this option to true.
+   * Other uses of the token do not consume it.
+   *
+   * This replay protection feature requires an additional network call to the App Check backend and forces the clients
+   * to obtain a fresh attestation from the chosen attestation providers. This can therefore negatively impact
+   * performance and can potentially deplete your attestation providers' quotas faster. Use this feature only for
+   * protecting low volume, security critical, or expensive operations.
+   *
+   * This option does not affect the enforceAppCheck option. Setting the latter to true will cause the callable function
+   * to automatically respond with a 401 Unauthorized status code when request includes an invalid App Check token.
+   * When request includes valid but consumed App Check tokens, requests will not be automatically rejected. Instead,
+   * the request.app.alreadyConsumed property will be set to true and pass the execution to the handler code for making
+   * further decisions, such as requiring additional security checks or rejecting the request.
+   */
+  consumeAppCheckToken?: boolean;
+
+  /**
    * Controls whether function configuration modified outside of function source is preserved. Defaults to false.
    *
    * @remarks
@@ -270,7 +294,7 @@ export interface DeploymentOptions extends RuntimeOptions {
   /**
    * Regions where function should be deployed.
    */
-  regions?: Array<(typeof SUPPORTED_REGIONS)[number] | string>;
+  regions?: Array<(typeof SUPPORTED_REGIONS)[number] | string | Expression<string> | ResetValue>;
   /**
    * Schedule for the scheduled function.
    */