+++ src/v2/providers/alerts/alerts.ts
@@ -70,7 +70,7 @@ export type AlertType =
   | "crashlytics.velocity"
   | "crashlytics.newAnrIssue"
   | "billing.planUpdate"
-  | "billing.automatedPlanUpdate"
+  | "billing.planAutomatedUpdate"
   | "appDistribution.newTesterIosDevice"
   | "appDistribution.inAppFeedback"
   | "performance.threshold"
@@ -94,7 +94,7 @@ export interface FirebaseAlertOptions extends options.EventHandlerOptions {
   /**
    * Region where functions should be deployed.
    */
-  region?: options.SupportedRegion | string;
+  region?: options.SupportedRegion | string | Expression<string> | ResetValue;
 
   /**
    * Amount of memory to allocate to a function.
@@ -163,7 +163,7 @@ export interface FirebaseAlertOptions extends options.EventHandlerOptions {
    * Specific service account for the function to run as.
    * A value of null restores the default service account.
    */
-  serviceAccount?: string | ResetValue;
+  serviceAccount?: string | Expression<string> | ResetValue;
 
   /**
    * Ingress settings which control where this function can be called from.