+++ src/v2/providers/database.ts
@@ -106,7 +106,7 @@ export interface ReferenceOptions<Ref extends string = string> extends options.E
   /**
    * Region where functions should be deployed.
    */
-  region?: options.SupportedRegion | string;
+  region?: options.SupportedRegion | string | Expression<string> | ResetValue;
 
   /**
    * Amount of memory to allocate to a function.
@@ -176,7 +176,7 @@ export interface ReferenceOptions<Ref extends string = string> extends options.E
   /**
    * Specific service account for the function to run as.
    */
-  serviceAccount?: string | ResetValue;
+  serviceAccount?: string | Expression<string> | ResetValue;
 
   /**
    * Ingress settings which control where this function can be called from.