+++ src/v2/providers/firestore.ts
@@ -35,6 +35,8 @@ import {
   createSnapshotFromProtobuf,
 } from "../../common/providers/firestore";
 
+export { Change };
+
 /** @internal */
 export const writtenEventType = "google.cloud.firestore.document.v1.written";
 
@@ -64,7 +66,7 @@ export interface RawFirestoreData {
 }
 
 /** @internal */
-export interface RawFirestoreEvent extends CloudEvent<Uint8Array | RawFirestoreData> {
+export interface RawFirestoreEvent extends CloudEvent<Uint8Array | RawFirestoreData | undefined> {
   location: string;
   project: string;
   database: string;
@@ -72,7 +74,6 @@ export interface RawFirestoreEvent extends CloudEvent<Uint8Array | RawFirestoreD
   document: string;
   datacontenttype?: string;
   dataschema: string;
-  time: string;
 }
 
 /** A Firestore DocumentSnapshot */
@@ -119,9 +120,9 @@ export interface DocumentOptions<Document extends string = string> extends Event
 export function onDocumentWritten<Document extends string>(
   document: Document,
   handler: (
-    event: FirestoreEvent<Change<DocumentSnapshot>, ParamsOf<Document>>
+    event: FirestoreEvent<Change<DocumentSnapshot> | undefined, ParamsOf<Document>>
   ) => any | Promise<any>
-): CloudFunction<FirestoreEvent<Change<DocumentSnapshot>, ParamsOf<Document>>>;
+): CloudFunction<FirestoreEvent<Change<DocumentSnapshot> | undefined, ParamsOf<Document>>>;
 
 /**
  * Event handler which triggers when a document is created, updated, or deleted in Firestore.
@@ -132,9 +133,9 @@ export function onDocumentWritten<Document extends string>(
 export function onDocumentWritten<Document extends string>(
   opts: DocumentOptions<Document>,
   handler: (
-    event: FirestoreEvent<Change<DocumentSnapshot>, ParamsOf<Document>>
+    event: FirestoreEvent<Change<DocumentSnapshot> | undefined, ParamsOf<Document>>
   ) => any | Promise<any>
-): CloudFunction<FirestoreEvent<Change<DocumentSnapshot>, ParamsOf<Document>>>;
+): CloudFunction<FirestoreEvent<Change<DocumentSnapshot> | undefined, ParamsOf<Document>>>;
 
 /**
  * Event handler which triggers when a document is created, updated, or deleted in Firestore.
@@ -145,9 +146,9 @@ export function onDocumentWritten<Document extends string>(
 export function onDocumentWritten<Document extends string>(
   documentOrOpts: Document | DocumentOptions<Document>,
   handler: (
-    event: FirestoreEvent<Change<DocumentSnapshot>, ParamsOf<Document>>
+    event: FirestoreEvent<Change<DocumentSnapshot> | undefined, ParamsOf<Document>>
   ) => any | Promise<any>
-): CloudFunction<FirestoreEvent<Change<DocumentSnapshot>, ParamsOf<Document>>> {
+): CloudFunction<FirestoreEvent<Change<DocumentSnapshot> | undefined, ParamsOf<Document>>> {
   return onChangedOperation(writtenEventType, documentOrOpts, handler);
 }
 
@@ -159,8 +160,10 @@ export function onDocumentWritten<Document extends string>(
  */
 export function onDocumentCreated<Document extends string>(
   document: Document,
-  handler: (event: FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>) => any | Promise<any>
-): CloudFunction<FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>>;
+  handler: (
+    event: FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>
+  ) => any | Promise<any>
+): CloudFunction<FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>>;
 
 /**
  * Event handler which triggers when a document is created in Firestore.
@@ -170,8 +173,10 @@ export function onDocumentCreated<Document extends string>(
  */
 export function onDocumentCreated<Document extends string>(
   opts: DocumentOptions<Document>,
-  handler: (event: FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>) => any | Promise<any>
-): CloudFunction<FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>>;
+  handler: (
+    event: FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>
+  ) => any | Promise<any>
+): CloudFunction<FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>>;
 
 /**
  * Event handler which triggers when a document is created in Firestore.
@@ -181,8 +186,10 @@ export function onDocumentCreated<Document extends string>(
  */
 export function onDocumentCreated<Document extends string>(
   documentOrOpts: Document | DocumentOptions<Document>,
-  handler: (event: FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>) => any | Promise<any>
-): CloudFunction<FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>> {
+  handler: (
+    event: FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>
+  ) => any | Promise<any>
+): CloudFunction<FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>> {
   return onOperation(createdEventType, documentOrOpts, handler);
 }
 
@@ -195,9 +202,9 @@ export function onDocumentCreated<Document extends string>(
 export function onDocumentUpdated<Document extends string>(
   document: Document,
   handler: (
-    event: FirestoreEvent<Change<QueryDocumentSnapshot>, ParamsOf<Document>>
+    event: FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, ParamsOf<Document>>
   ) => any | Promise<any>
-): CloudFunction<FirestoreEvent<Change<QueryDocumentSnapshot>, ParamsOf<Document>>>;
+): CloudFunction<FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, ParamsOf<Document>>>;
 /**
  * Event handler which triggers when a document is updated in Firestore.
  *
@@ -207,9 +214,9 @@ export function onDocumentUpdated<Document extends string>(
 export function onDocumentUpdated<Document extends string>(
   opts: DocumentOptions<Document>,
   handler: (
-    event: FirestoreEvent<Change<QueryDocumentSnapshot>, ParamsOf<Document>>
+    event: FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, ParamsOf<Document>>
   ) => any | Promise<any>
-): CloudFunction<FirestoreEvent<Change<QueryDocumentSnapshot>, ParamsOf<Document>>>;
+): CloudFunction<FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, ParamsOf<Document>>>;
 
 /**
  * Event handler which triggers when a document is updated in Firestore.
@@ -220,9 +227,9 @@ export function onDocumentUpdated<Document extends string>(
 export function onDocumentUpdated<Document extends string>(
   documentOrOpts: Document | DocumentOptions<Document>,
   handler: (
-    event: FirestoreEvent<Change<QueryDocumentSnapshot>, ParamsOf<Document>>
+    event: FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, ParamsOf<Document>>
   ) => any | Promise<any>
-): CloudFunction<FirestoreEvent<Change<QueryDocumentSnapshot>, ParamsOf<Document>>> {
+): CloudFunction<FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, ParamsOf<Document>>> {
   return onChangedOperation(updatedEventType, documentOrOpts, handler);
 }
 
@@ -234,8 +241,10 @@ export function onDocumentUpdated<Document extends string>(
  */
 export function onDocumentDeleted<Document extends string>(
   document: Document,
-  handler: (event: FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>) => any | Promise<any>
-): CloudFunction<FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>>;
+  handler: (
+    event: FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>
+  ) => any | Promise<any>
+): CloudFunction<FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>>;
 
 /**
  * Event handler which triggers when a document is deleted in Firestore.
@@ -245,8 +254,10 @@ export function onDocumentDeleted<Document extends string>(
  */
 export function onDocumentDeleted<Document extends string>(
   opts: DocumentOptions<Document>,
-  handler: (event: FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>) => any | Promise<any>
-): CloudFunction<FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>>;
+  handler: (
+    event: FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>
+  ) => any | Promise<any>
+): CloudFunction<FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>>;
 
 /**
  * Event handler which triggers when a document is deleted in Firestore.
@@ -256,8 +267,10 @@ export function onDocumentDeleted<Document extends string>(
  */
 export function onDocumentDeleted<Document extends string>(
   documentOrOpts: Document | DocumentOptions<Document>,
-  handler: (event: FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>) => any | Promise<any>
-): CloudFunction<FirestoreEvent<QueryDocumentSnapshot, ParamsOf<Document>>> {
+  handler: (
+    event: FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>
+  ) => any | Promise<any>
+): CloudFunction<FirestoreEvent<QueryDocumentSnapshot | undefined, ParamsOf<Document>>> {
   return onOperation(deletedEventType, documentOrOpts, handler);
 }
 
@@ -345,20 +358,17 @@ export function makeFirestoreEvent<Params>(
   eventType: string,
   event: RawFirestoreEvent,
   params: Params
-): FirestoreEvent<QueryDocumentSnapshot, Params> {
+): FirestoreEvent<QueryDocumentSnapshot | undefined, Params> {
   const data = event.data
     ? eventType === createdEventType
       ? createSnapshot(event)
       : createBeforeSnapshot(event)
     : undefined;
-  const firestoreEvent: FirestoreEvent<QueryDocumentSnapshot, Params> = {
+  const firestoreEvent: FirestoreEvent<QueryDocumentSnapshot | undefined, Params> = {
     ...event,
     params,
     data,
   };
-  if (firestoreEvent.data === undefined) {
-    delete (firestoreEvent as any).data;
-  }
   delete (firestoreEvent as any).datacontenttype;
   delete (firestoreEvent as any).dataschema;
   return firestoreEvent;
@@ -368,18 +378,15 @@ export function makeFirestoreEvent<Params>(
 export function makeChangedFirestoreEvent<Params>(
   event: RawFirestoreEvent,
   params: Params
-): FirestoreEvent<Change<QueryDocumentSnapshot>, Params> {
+): FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, Params> {
   const data = event.data
     ? Change.fromObjects(createBeforeSnapshot(event), createSnapshot(event))
     : undefined;
-  const firestoreEvent: FirestoreEvent<Change<QueryDocumentSnapshot>, Params> = {
+  const firestoreEvent: FirestoreEvent<Change<QueryDocumentSnapshot> | undefined, Params> = {
     ...event,
     params,
     data,
   };
-  if (firestoreEvent.data === undefined) {
-    delete (firestoreEvent as any).data;
-  }
   delete (firestoreEvent as any).datacontenttype;
   delete (firestoreEvent as any).dataschema;
   return firestoreEvent;